requirements.txt
asyncio
langchain-community
langchain-google-genai
google-generativeai
python-dotenv